Extend Submission Requirement for Arbitrary Logic
These changes allow Submission Requirement to support arbitrary logic to
support the use case described in the comment here:

https://github.com/decentralized-identity/presentation-exchange/pull/18/files#diff-9308a2575ae14de31b25f4459ecfc604R57

Note that while `pick` can require "exactly three" we do not currently
support a rule that allows for "three or more", but this could be added
in the future.

Added a test to validate that an instance in the new format satisfying
the use case above also adheres to the JSON Schema.